# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: WB Analytics API
  version: 1.0.0
  description: Контракт API сервиса аналитики товаров Wildberries. Снят с Django-реализации
    и заморожен на время миграции.
paths:
...
  /api/schedules/:
    get:
      operationId: schedules_list
      summary: List the caller's schedules
      tags:
      - schedules
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schedule'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
    post:
      operationId: schedules_create
      summary: Create a schedule
      tags:
      - schedules
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ScheduleRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ScheduleRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
  /api/schedules/{schedule_id}/:
    patch:
      operationId: schedules_set_active
      summary: Enable or disable a schedule
      parameters:
      - in: path
        name: schedule_id
        schema:
          type: integer
        required: true
      tags:
      - schedules
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedScheduleRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedScheduleRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedScheduleRequest'
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
    delete:
      operationId: schedules_destroy
      summary: Delete a schedule
      parameters:
      - in: path
        name: schedule_id
        schema:
          type: integer
        required: true
      tags:
      - schedules
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '204':
          description: No response body
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
...
components:
  schemas:
...
    Error:
      type: object
      description: |-
        The `{"detail": ...}` body the shared exception handler returns.

        Not the only error body in the API: DRF answers its own serializer
        validation with a field-keyed map instead, and the handler passes that
        through untouched. The docstring here used to claim this shape covered
        "every error path", which is what a contract test disproved — see
        `FIELD_ERRORS_SCHEMA` below and `VALIDATION_ERROR_RESPONSE`.
      properties:
        detail:
          type: string
      required:
      - detail
...
    PatchedScheduleRequest:
      type: object
      properties:
        query:
          type: string
          minLength: 1
        spec:
          type: string
          minLength: 1
        active:
          type: boolean
          default: true
...
    Schedule:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        query:
          type: string
        spec:
          type: string
        interval_seconds:
          type: integer
          readOnly: true
        active:
          type: boolean
          default: true
        last_run_at:
          type:
          - string
          - 'null'
          format: date-time
          readOnly: true
      required:
      - id
      - interval_seconds
      - last_run_at
      - query
      - spec
    ScheduleRequest:
      type: object
      properties:
        query:
          type: string
          minLength: 1
        spec:
          type: string
          minLength: 1
        active:
          type: boolean
          default: true
      required:
      - query
      - spec
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT